// types.d.ts
type mixed = unknown;
export type MayBe<T> = T | null | undefined; 
export type LangFunction = (s: string) => string;
// ...
type RecursiveArray<T> =
  | null
  | void
  | T
  | false
  | ''
  | ReadonlyArray<RecursiveArray<T>>;
export type StyleProp = RecursiveArray<Readonly<{[prop: string]: mixed}>>;

export interface Ref {
    clearValidation(): void,
    readonly isValid: boolean,
    validate(): boolean,
    validateAsync(): Promise<boolean>,
}

export interface ContextRef extends Ref {
    getErrorMessage(name: string): string | void,
    getInput(name: string): InputRef | void,
    refreshMessage(): void,
    setErrorMessage(name: string, message: string): void,
}

export interface InputRef extends Ref {
    focus?: () => mixed,
    getErrorMessage: () => string,
    readonly name?: string,
    setErrorMessage(message: string): void,
}

export type RefObject<T> = {current: T | null, [prop: string]: mixed};
export type RefProp<T> = ((ref: T | null) => mixed) | RefObject<T>;

type TAsyncFailMessage = {
    Default: 0,
    CaughtError: 1,
};
export declare const AsyncFailMessage: TAsyncFailMessage;

export type ContextValue = {
    asyncFailMessage: TAsyncFailMessage[keyof TAsyncFailMessage],
    auto: boolean,
    errorTextStyle: StyleProp,
    focusOnInvalid: boolean,
    inputErrorStyle: StyleProp,
    lang?: LangFunction,
    addRef: (ref: InputRef) => mixed,
    removeRef: (ref: InputRef) => mixed;
};

export type ContextProps = {
    asyncFailMessage: ContextValue['asyncFailMessage'],
    auto: ContextValue['auto'],
    children: React.ReactNode,
    errorTextStyle: ContextValue['errorTextStyle'],
    focusOnInvalid: ContextValue['focusOnInvalid'],
    inputErrorStyle: ContextValue['inputErrorStyle'],
    lang: NonNullable<ContextValue['lang']>,
};

export type ContextDefaultProps = {
    asyncFailMessage: ContextProps['asyncFailMessage'],
    auto: ContextProps['auto'],
    errorTextStyle: ContextProps['errorTextStyle'],
    focusOnInvalid: ContextProps['focusOnInvalid'],
    inputErrorStyle: ContextProps['inputErrorStyle'],
    lang: ContextProps['lang'],
};